+++ CHANGELOG.md
@@ -5,8 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 
-## 0.15.17
+## 0.15.18
 - Implement pahination for predictive data MCP tools
+
+## 0.15.17
+- Added option to configure OAuth2 token exchange flow for server
 
 ## 0.15.16
 - Pinned ag-ui-protocol to version 0.1.15